Rename auth module and remove all testing code
