--- ui/src/app/core/services/blog/blog.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { BlogService } from './blog.service';
-
-describe('BlogService', () => {
-    let service: BlogService;
-
-    beforeEach(() => {
-        TestBed.configureTestingModule({});
-        service = TestBed.inject(BlogService);
-    });
-
-    it('should be created', () => {
-        expect(service).toBeTruthy();
-    });
-});
