+++ ui/src/app/core/services/comparison.service.ts
