--- ui/src/app/core/services/comparison/comparison.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ComparisonService } from './comparison.service';
-
-describe('ComparisonService', () => {
-    let service: ComparisonService;
-
-    beforeEach(() => {
-        TestBed.configureTestingModule({});
-        service = TestBed.inject(ComparisonService);
-    });
-
-    it('should be created', () => {
-        expect(service).toBeTruthy();
-    });
-});
