+++ ui/src/app/core/services/editor.service.ts
