--- ui/src/app/core/services/editor/editor.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { EditorService } from './editor.service';
-
-describe('EditorService', () => {
-    let service: EditorService;
-
-    beforeEach(() => {
-        TestBed.configureTestingModule({});
-        service = TestBed.inject(EditorService);
-    });
-
-    it('should be created', () => {
-        expect(service).toBeTruthy();
-    });
-});
