--- ui/src/app/core/services/notification/notification.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { NotificationService } from './notification.service';
-
-describe('NotificationService', () => {
-    let service: NotificationService;
-
-    beforeEach(() => {
-        TestBed.configureTestingModule({});
-        service = TestBed.inject(NotificationService);
-    });
-
-    it('should be created', () => {
-        expect(service).toBeTruthy();
-    });
-});
