+++ ui/src/app/core/services/validation.service.ts
