--- ui/src/app/modules/editor/components/post-editor/post-editor.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { PostEditorComponent } from './post-editor.component';
-
-describe('PostEditorComponent', () => {
-    let component: PostEditorComponent;
-    let fixture: ComponentFixture<PostEditorComponent>;
-
-    beforeEach(async(() => {
-        TestBed.configureTestingModule({
-            declarations: [PostEditorComponent]
-        })
-            .compileComponents();
-    }));
-
-    beforeEach(() => {
-        fixture = TestBed.createComponent(PostEditorComponent);
-        component = fixture.componentInstance;
-        fixture.detectChanges();
-    });
-
-    it('should create', () => {
-        expect(component).toBeTruthy();
-    });
-});
