--- ui/src/app/modules/editor/components/profile-editor/profile-editor.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ProfileEditorComponent } from './profile-editor.component';
-
-describe('ProfileEditorComponent', () => {
-    let component: ProfileEditorComponent;
-    let fixture: ComponentFixture<ProfileEditorComponent>;
-
-    beforeEach(async(() => {
-        TestBed.configureTestingModule({
-            declarations: [ProfileEditorComponent]
-        })
-            .compileComponents();
-    }));
-
-    beforeEach(() => {
-        fixture = TestBed.createComponent(ProfileEditorComponent);
-        component = fixture.componentInstance;
-        fixture.detectChanges();
-    });
-
-    it('should create', () => {
-        expect(component).toBeTruthy();
-    });
-});
