--- ui/src/app/modules/editor/components/project-editor/project-editor.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ProjectEditorComponent } from './project-editor.component';
-
-describe('ProjectEditorComponent', () => {
-    let component: ProjectEditorComponent;
-    let fixture: ComponentFixture<ProjectEditorComponent>;
-
-    beforeEach(async(() => {
-        TestBed.configureTestingModule({
-            declarations: [ProjectEditorComponent]
-        })
-            .compileComponents();
-    }));
-
-    beforeEach(() => {
-        fixture = TestBed.createComponent(ProjectEditorComponent);
-        component = fixture.componentInstance;
-        fixture.detectChanges();
-    });
-
-    it('should create', () => {
-        expect(component).toBeTruthy();
-    });
-});
