--- ui/src/app/modules/icon/components/github-icon/github-icon.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { GithubIconComponent } from './github-icon.component';
-
-describe('GithubIconComponent', () => {
-    let component: GithubIconComponent;
-    let fixture: ComponentFixture<GithubIconComponent>;
-
-    beforeEach(async(() => {
-        TestBed.configureTestingModule({
-            declarations: [GithubIconComponent]
-        })
-            .compileComponents();
-    }));
-
-    beforeEach(() => {
-        fixture = TestBed.createComponent(GithubIconComponent);
-        component = fixture.componentInstance;
-        fixture.detectChanges();
-    });
-
-    it('should create', () => {
-        expect(component).toBeTruthy();
-    });
-});
