--- ui/src/app/shared/components/post-collection/post-collection.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { PostCollectionComponent } from './post-collection.component';
-
-describe('PostCollectionComponent', () => {
-    let component: PostCollectionComponent;
-    let fixture: ComponentFixture<PostCollectionComponent>;
-
-    beforeEach(async(() => {
-        TestBed.configureTestingModule({
-            declarations: [PostCollectionComponent]
-        })
-            .compileComponents();
-    }));
-
-    beforeEach(() => {
-        fixture = TestBed.createComponent(PostCollectionComponent);
-        component = fixture.componentInstance;
-        fixture.detectChanges();
-    });
-
-    it('should create', () => {
-        expect(component).toBeTruthy();
-    });
-});
