--- ui/src/app/shared/components/project-collection/project-collection.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ProjectCollectionComponent } from './project-collection.component';
-
-describe('ProjectCollectionComponent', () => {
-    let component: ProjectCollectionComponent;
-    let fixture: ComponentFixture<ProjectCollectionComponent>;
-
-    beforeEach(async(() => {
-        TestBed.configureTestingModule({
-            declarations: [ProjectCollectionComponent]
-        })
-            .compileComponents();
-    }));
-
-    beforeEach(() => {
-        fixture = TestBed.createComponent(ProjectCollectionComponent);
-        component = fixture.componentInstance;
-        fixture.detectChanges();
-    });
-
-    it('should create', () => {
-        expect(component).toBeTruthy();
-    });
-});
